Add Transaction::ReadIfExists interface
The new Transaction::ReadIfExists provides the functionality of the
Transaction::Read operation but does not throw an exception when the
requested object doesn't exist.  Instead, a return flag is used to
indicate whether or not that requested object exists.  Eventually,
this will give the caller the option to avoid incurring the cost of
handling the ObjectDoesntExistException.  At the moment, the cost is
only avoided for batched reads.